gridsight_dashboard/requirements.txt
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
azure-storage-blob>=12.17.0
python-dateutil>=2.8.2 